Added similarity search to correlate results.

+++ sherwood-server/src/main/java/io/steviemul/sherwood/server/response/SarifResultSimilarityResponse.java
@@ -5,7 +5,7 @@ import java.util.UUID;
 public record SarifResultSimilarityResponse(
     UUID matchingResultId,
     String location,
-    int lineNumber,
+    long lineNumber,
     String ruleId,
     double similarity,
     String reason) {}
